refaktorisering av daterange komponenten
